useHandleDrag hook
