+++ apps/desktop/src/renderer/components/App/App.tsx
@@ -1,7 +1,6 @@
-// import { Audio } from '../Audio'
-import React, { useRef, useState } from 'react'
 import { AppStoreProvider, EngineStoreProvider, WidgetStrip, useAppStore } from '@hedron-gl/ui-core'
 import c from './App.module.css'
+import { useHandleDrag } from './useHandleDrag'
 import { GlobalClock } from '@components/GlobalClock/GlobalClock'
 import { GlobalDialogs } from '@components/GlobalDialogs/GlobalDialogs'
 import { PerformanceStats } from '@components/PerformanceStats/PerformanceStats'
@@ -11,46 +10,14 @@ import { WorkArea } from '@components/WorkArea/WorkArea'
 import { appStore } from '@renderer/appStore'
 import { engine, engineStore, pluginViews } from '@renderer/engine'
 
-const MIN_RATIO = 0.25
-const MAX_RATIO = 0.75
-
 const AppContent = (): JSX.Element => {
   const sketchesDir = useAppStore((state) => state.sketchesDir)
   const isProjectLoaded = sketchesDir !== null
-  const [leftRatio, setLeftRatio] = useState(0.5)
-  const dragging = useRef(false)
 
-  const onMouseDown = () => {
-    dragging.current = true
-    document.body.style.cursor = 'col-resize'
-  }
-
-  React.useEffect(() => {
-    const onMouseMove = (e: MouseEvent) => {
-      if (!dragging.current) return
-      const wrapper = document.getElementById('app-wrapper')
-      if (!wrapper) return
-      const rect = wrapper.getBoundingClientRect()
-      let ratio = (e.clientX - rect.left) / rect.width
-      ratio = Math.max(MIN_RATIO, Math.min(MAX_RATIO, ratio))
-      setLeftRatio(ratio)
-    }
-    const onMouseUp = () => {
-      if (dragging.current) {
-        dragging.current = false
-        document.body.style.cursor = ''
-      }
-    }
-    window.addEventListener('mousemove', onMouseMove)
-    window.addEventListener('mouseup', onMouseUp)
-    return () => {
-      window.removeEventListener('mousemove', onMouseMove)
-      window.removeEventListener('mouseup', onMouseUp)
-    }
-  }, [])
+  const { leftRatio, onHandleMouseDown, wrapperRef } = useHandleDrag()
 
   return (
-    <div className={c.wrapper} id="app-wrapper">
+    <div className={c.wrapper} ref={wrapperRef}>
       <div
         className={c.left}
         style={{ flexBasis: `${leftRatio * 100}%`, maxWidth: `${leftRatio * 100}%` }}
@@ -76,7 +43,7 @@ const AppContent = (): JSX.Element => {
       </div>
       <div
         className={c.handle}
-        onMouseDown={onMouseDown}
+        onMouseDown={onHandleMouseDown}
         role="separator"
         aria-orientation="vertical"
         tabIndex={0}
