+++ packages/ui-core/src/index.ts
@@ -30,3 +30,4 @@ export * from './hooks/useUpdateNodeValue'
 export * from './hooks/useSubscribeToShot'
 export * from './hooks/useElementScrub'
 export * from './components/WidgetStrip'
+export * from './utils/setGlobalCursor'
